Moving everything regarding platform-specifics to map_platform_specifics.py
makescript.py is now essentially redundant, but we can keep it around for a while and just add a warning, and make it pass the call to map_platform_specifics.py.

pyproject.toml is now handled by having a "skeleton" __pyproject.toml, which is parsed by set_toml_version(), which sets the appropriate version number based on the --diffs flag and writes the actual pyproject.toml.